fix select box in checkout page

+++ database/migrations/2023_05_08_025411_create_products_table.php
@@ -19,7 +19,7 @@ class CreateProductsTable extends Migration
             $table->string('name');
             $table->string('image');
             $table->longText('description');
-            $table->decimal('price', 20, 2);
+            $table->integer('price');
             $table->timestamps();
         });
     }
